// frontend/src/features/students/types/student.types.ts
export interface AdminStudentListItem {
  id: string;
  email: string;
  displayName?: string | null;
  isVerified: boolean;
  isActive: boolean;
  fullName: string;
  studentCode: string;
  faculty: string;
  major?: string | null;
  administrativeClass: string;
  academicYear: number;
  school: string;
  createdAt: string;
}